reorganized the code for ENM classes
